fix small issues

+++ modules/Nncase.Modules.K210/Transform/Rules/K210/RealizeFakeDequantize.cs
@@ -42,6 +42,6 @@ public sealed partial class RealizeFakeDequantize : IRewriteRule
         var qm = QuantMode.UnsignedMode;
         var qp = GetQuantParam((range[0], range[1]), bits, qm);
         qp.Scale = 1f / qp.Scale;
-        return IR.F.Math.Dequantize(input, qp, DataTypes.UInt8);
+        return IR.F.Math.Dequantize(input, qp, DataTypes.Float32);
     }
 }
